need work on the squigly lines stuck on
